src/requirements.txt
torch==2.6.0
transformers==4.46.3
tokenizers==0.20.3
tqdm
einops
addict 
easydict
PyQt6
zstandard
numpy
pandas
Pillow
pycryptodome
pymupdf
img2pdf
reportlab